refactor(examples): make examples more consistent
