+++ examples/with-jest-jsdom/tests/example.test.ts
@@ -6,38 +6,41 @@ import renderApp, { GitHubRepository } from '../src/app';
 import githubInterceptor from './interceptors/github';
 
 describe('Example tests', () => {
+  const ownerName = 'diego-aquino';
+  const repositoryName = 'zimic';
+
+  const repository: GitHubRepository = {
+    id: 1,
+    full_name: 'diego-aquino/zimic',
+    html_url: 'https://github.com/diego-aquino/zimic',
+  };
+
   beforeEach(() => {
     document.body.innerHTML = '';
   });
 
   it('should render a GitHub repository, if found', async () => {
-    const zimicRepository: GitHubRepository = {
-      id: 1,
-      full_name: 'diego-aquino/zimic',
-      html_url: 'https://github.com/diego-aquino/zimic',
-    };
-
-    const getRepositoryHandler = githubInterceptor.get('/repos/:owner/:name').respond({
+    const getRepositoryHandler = githubInterceptor.get(`/repos/${ownerName}/${repositoryName}`).respond({
       status: 200,
-      body: zimicRepository,
+      body: repository,
     });
 
     renderApp();
 
     const ownerInput = screen.getByRole('textbox', { name: 'Owner' });
-    await userEvent.type(ownerInput, 'diego-aquino');
+    await userEvent.type(ownerInput, ownerName);
 
     const repositoryInput = screen.getByRole('textbox', { name: 'Repository' });
-    await userEvent.type(repositoryInput, 'zimic');
+    await userEvent.type(repositoryInput, repositoryName);
 
     const searchButton = screen.getByRole('button', { name: 'Search' });
     await userEvent.click(searchButton);
 
-    const repositoryName = await screen.findByRole('heading', { name: zimicRepository.full_name });
-    expect(repositoryName).toBeInTheDocument();
+    const repositoryHeading = await screen.findByRole('heading', { name: repository.full_name });
+    expect(repositoryHeading).toBeInTheDocument();
 
-    const repositoryURL = await screen.findByRole('link', { name: zimicRepository.html_url });
-    expect(repositoryURL).toBeInTheDocument();
+    const repositoryLink = await screen.findByRole('link', { name: repository.html_url });
+    expect(repositoryLink).toBeInTheDocument();
 
     const getRequests = getRepositoryHandler.requests();
     expect(getRequests).toHaveLength(1);
@@ -52,19 +55,19 @@ describe('Example tests', () => {
     renderApp();
 
     const ownerInput = screen.getByRole('textbox', { name: 'Owner' });
-    await userEvent.type(ownerInput, 'diego-aquino');
+    await userEvent.type(ownerInput, ownerName);
 
     const repositoryInput = screen.getByRole('textbox', { name: 'Repository' });
-    await userEvent.type(repositoryInput, 'zimic');
+    await userEvent.type(repositoryInput, repositoryName);
 
     const searchButton = screen.getByRole('button', { name: 'Search' });
     await userEvent.click(searchButton);
 
-    const repositoryName = await screen.findByRole('heading', { name: 'Repository not found' });
-    expect(repositoryName).toBeInTheDocument();
+    const repositoryHeading = await screen.findByRole('heading', { name: 'Repository not found' });
+    expect(repositoryHeading).toBeInTheDocument();
 
-    const repositoryURLPromise = screen.findByRole('link', {}, { timeout: 200 });
-    await expect(repositoryURLPromise).rejects.toThrowError();
+    const repositoryLinkPromise = screen.findByRole('link', {}, { timeout: 200 });
+    await expect(repositoryLinkPromise).rejects.toThrowError();
 
     const getRequests = getRepositoryHandler.requests();
     expect(getRequests).toHaveLength(1);
