+++ examples/with-vitest-node/tests/example.test.ts
@@ -5,6 +5,15 @@ import app, { GitHubRepository } from '../src/app';
 import githubInterceptor from './interceptors/github';
 
 describe('Example tests', () => {
+  const ownerName = 'diego-aquino';
+  const repositoryName = 'zimic';
+
+  const repository: GitHubRepository = {
+    id: 1,
+    full_name: 'diego-aquino/zimic',
+    html_url: 'https://github.com/diego-aquino/zimic',
+  };
+
   beforeAll(async () => {
     await app.ready();
   });
@@ -14,23 +23,17 @@ describe('Example tests', () => {
   });
 
   it('should return a GitHub repository, if found', async () => {
-    const zimicRepository: GitHubRepository = {
-      id: 1,
-      full_name: 'diego-aquino/zimic',
-      html_url: 'https://github.com/diego-aquino/zimic',
-    };
-
-    const getRepositoryHandler = githubInterceptor.get('/repos/:owner/:name').respond({
+    const getRepositoryHandler = githubInterceptor.get(`/repos/${ownerName}/${repositoryName}`).respond({
       status: 200,
-      body: zimicRepository,
+      body: repository,
     });
 
-    const response = await supertest(app.server).get('/github/repositories/diego-aquino/zimic');
+    const response = await supertest(app.server).get(`/github/repositories/${ownerName}/${repositoryName}`);
     expect(response.status).toBe(200);
     expect(response.body).toEqual({
-      id: zimicRepository.id,
-      fullName: zimicRepository.full_name,
-      homepageURL: zimicRepository.html_url,
+      id: repository.id,
+      fullName: repository.full_name,
+      homepageURL: repository.html_url,
     });
 
     const getRequests = getRepositoryHandler.requests();
@@ -43,7 +46,7 @@ describe('Example tests', () => {
       body: { message: 'Not Found' },
     });
 
-    const response = await supertest(app.server).get('/github/repositories/diego-aquino/zimic');
+    const response = await supertest(app.server).get(`/github/repositories/${ownerName}/${repositoryName}`);
     expect(response.status).toBe(404);
     expect(response.body).toEqual({});
